+++ Package.swift
@@ -10,8 +10,8 @@ let package = Package(
     targets: [
         .binaryTarget(
             name: "QvpEngine",
-            url: "https://github.com/quran-ws/quran-engine/releases/download/v0.2.1/QvpEngine.xcframework.zip",
-            checksum: "9490ec567790874295a3cb611472f4185dbb43ca76bf360e53d13fec460f29de"
+            url: "https://github.com/quran-ws/quran-engine/releases/download/v0.2.2/QvpEngine.xcframework.zip",
+            checksum: "8ce6d7b4ccf1c8dad896b2a82a9159b2041965c1e296ad2d0d3f5f11c8b4436a"
         ),
         .target(
             name: "QvpKit",